requirements.txt
numpy
pandas
dash
yfinance
plotly.express
plotly.graph_objects